[Refactor] userId 타입 Integer -> Long으로 변환 #73 Related to: #38, #18

+++ src/main/java/com/midnear/midnearshopping/mapper/coupon_point/UserCouponMapper.java
@@ -1,14 +1,13 @@
 package com.midnear.midnearshopping.mapper.coupon_point;
 
 import com.midnear.midnearshopping.domain.dto.coupon_point.CouponInfoDto;
-import com.midnear.midnearshopping.domain.vo.coupon_point.UserCouponVO;
 import org.apache.ibatis.annotations.Mapper;
 
 import java.util.List;
 
 @Mapper
 public interface UserCouponMapper {
-    List<CouponInfoDto> getCouponsByUserId(Integer userId);
-    int getCouponsCountByUserId(Integer userId);
+    List<CouponInfoDto> getCouponsByUserId(Long userId);
+    int getCouponsCountByUserId(Long userId);
     void changeStatus(Long userCouponId);
 }
